Move common test datatypes into `test_helpers.go`
