+++ taglocker_test.go
@@ -4,60 +4,8 @@ import (
 	"fmt"
 	"math/rand"
 	"strings"
-	"sync"
 	"testing"
 )
-
-const alphaCharset = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-const numCharset = "0123456789"
-const alphanumCharset = alphaCharset + "_" + numCharset
-
-// mockReadWriter is a dummy ReadWriter interface that unconditionally succeeds.  We
-// protect concurrenct accesses to the `state` map but _not_ the memory that the map's
-// value points to, so it has to be the responsibility of the caller of ReadTag and
-// WriteTag to ensure mutual exclusion on _their_ accesses.
-type mockReadWriter struct {
-	state map[string]*uint32 // State is read/written to for the benefit of the race detector during unit tests.
-	mtx   sync.Mutex
-}
-
-var _ = ReadWriter(&TagLocker{})
-
-func newMockReadWriter() *mockReadWriter {
-	return &mockReadWriter{
-		state: make(map[string]*uint32),
-	}
-}
-func (norw *mockReadWriter) ReadTag(name string, value interface{}) error {
-	//fmt.Fprintf(os.Stderr, "Downstream: Reading %v\n", name)
-
-	norw.mtx.Lock()
-	ptr, ok := norw.state[name]
-	if !ok {
-		ptr = new(uint32)
-		norw.state[name] = ptr
-	}
-	norw.mtx.Unlock()
-
-	value = *ptr
-
-	return nil
-}
-func (norw *mockReadWriter) WriteTag(name string, value interface{}) error {
-	//fmt.Fprintf(os.Stderr, "Downstream: Writing %v\n", name)
-
-	norw.mtx.Lock()
-	ptr, ok := norw.state[name]
-	if !ok {
-		ptr = new(uint32)
-		norw.state[name] = ptr
-	}
-	norw.mtx.Unlock()
-
-	*ptr++
-
-	return nil
-}
 
 // Generates a random fully-qualified tag, and returns
 // paths to that tag but also all intermediary prefixes.
@@ -118,47 +66,40 @@ func genPaths() []string {
 	return paths
 }
 
-const serialConcurrency = 1
-const lowConcurrency = 10
-const medConcurrency = 20
-const highConcurrency = 30
-const writePerc = 1
-const heavyWritePerc = 50
-
-const numTags = 10
-
 func BenchmarkSerialTagLocking(b *testing.B) {
-	benchmarkTagLockLocking(b, serialConcurrency, writePerc)
+	benchmarkTagLockLocking(b, serialTestConcurrency, testWritePrecentage)
 }
 
 func BenchmarkLowConcurrencyTagLocking(b *testing.B) {
-	benchmarkTagLockLocking(b, lowConcurrency, writePerc)
+	benchmarkTagLockLocking(b, lowTestConcurrency, testWritePrecentage)
 }
 
 func BenchmarkMedConcurrencyTagLocking(b *testing.B) {
-	benchmarkTagLockLocking(b, medConcurrency, writePerc)
+	benchmarkTagLockLocking(b, mediumTestConcurrency, testWritePrecentage)
 }
 
 func BenchmarkHighConcurrencyTagLocking(b *testing.B) {
-	benchmarkTagLockLocking(b, highConcurrency, writePerc)
+	benchmarkTagLockLocking(b, highTestConcurrency, testWritePrecentage)
 }
 
 func BenchmarkSerialTagLockingWithHeavyWrites(b *testing.B) {
-	benchmarkTagLockLocking(b, serialConcurrency, heavyWritePerc)
+	benchmarkTagLockLocking(b, serialTestConcurrency, heavyWritePerc)
 }
 
 func BenchmarkLowConcurrencyTagLockingWithHeavyWrites(b *testing.B) {
-	benchmarkTagLockLocking(b, lowConcurrency, heavyWritePerc)
+	benchmarkTagLockLocking(b, lowTestConcurrency, heavyWritePerc)
 }
 
 func BenchmarkMedConcurrencyTagLockingWithHeavyWrites(b *testing.B) {
-	benchmarkTagLockLocking(b, medConcurrency, heavyWritePerc)
+	benchmarkTagLockLocking(b, mediumTestConcurrency, heavyWritePerc)
 }
 func BenchmarkHighConcurrencyTagLockingWithHeavyWrites(b *testing.B) {
-	benchmarkTagLockLocking(b, highConcurrency, heavyWritePerc)
+	benchmarkTagLockLocking(b, highTestConcurrency, heavyWritePerc)
 }
 
 func benchmarkTagLockLocking(b *testing.B, concurrency int, writePerc int) {
+	const numTags = 10
+
 	barrier := make(chan bool, concurrency)
 
 	tl := NewTagLocker(newMockReadWriter())
